Unit tests, and Resource Optimization for IncludeDiagnostics option
Added a couple of unit tests to verify output of sink when using the
IncludeDiagnostics flag.  I also did a bit of naming cleanup, and made
the PackageContent function only allocate the new object and Json
Serialize it if includeDiagnostics is turned on to reduce the amount of
useless work the sink does.